requirements.txt
langchain
langchain-community
langchain-ollama
chromadb
unstructured[md]
python-docx
pypdf
gradio
watchdog
python-dotenv